+++ src/main/java/dandi/dandi/batchcommons/config/BatchConfig.java
@@ -1,4 +1,4 @@
-package dandi.dandi.batch.config;
+package dandi.dandi.batchcommons.config;
 
 import org.springframework.batch.core.configuration.annotation.EnableBatchProcessing;
 import org.springframework.context.annotation.Configuration;
