+++ src/test/java/dandi/dandi/unusedimagebatch/application/service/UnusedImageBatchSchedulerTest.java
@@ -1,6 +1,6 @@
 package dandi.dandi.unusedimagebatch.application.service;
 
-import dandi.dandi.batch.exception.BatchException;
+import dandi.dandi.batchcommons.exception.BatchException;
 import dandi.dandi.errormessage.application.port.out.ErrorMessageSender;
 import dandi.dandi.unusedimagebatch.application.runner.UnusedImageDeletionBatch;
 import org.junit.jupiter.api.DisplayName;
